requirements.txt
requests
beautifulsoup4==4.5.3
py4j==0.10.4
redis==2.10.5
kafka-python==1.3.3
nltk==3.2.2
Flask==0.12
Flask-Cors==3.0.2